Refactored LLM NPUW accuracy tests
