+++ src/plugins/intel_npu/tests/functional/behavior/npuw/test_engine/comparators/nrmse.hpp
