+++ src/plugins/intel_npu/tests/functional/behavior/npuw/test_engine/generators/model_generator.hpp
