+++ src/plugins/intel_npu/tests/functional/behavior/npuw/test_engine/mocks/register_in_ov.hpp
